#141 Test: move MainDispatcherRule to domain:testing
